+++ optuna_dashboard/_inmemory_cache.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import copy
 from datetime import datetime
-from datetime import timedelta
 import numbers
 import threading
 from typing import List
@@ -12,7 +11,6 @@ from typing import Tuple
 from typing import TYPE_CHECKING
 
 from optuna.distributions import BaseDistribution
-from optuna.storages import BaseStorage
 from optuna.trial import FrozenTrial
 from optuna.trial import TrialState
 
@@ -41,35 +39,6 @@ def get_cached_extra_study_property(
         )
 
 
-def get_trials(
-    in_memory_cache: InMemoryCache, storage: BaseStorage, study_id: int
-) -> list[FrozenTrial]:
-    with in_memory_cache._trials_cache_lock:
-        trials = in_memory_cache._trials_cache.get(study_id, None)
-
-        # Not a big fan of the heuristic, but I can't think of anything better.
-        if trials is None or len(trials) < 100:
-            ttl_seconds = 2
-        elif len(trials) < 500:
-            ttl_seconds = 5
-        else:
-            ttl_seconds = 10
-
-        last_fetched_at = in_memory_cache._trials_last_fetched_at.get(study_id, None)
-        if (
-            trials is not None
-            and last_fetched_at is not None
-            and datetime.now() - last_fetched_at < timedelta(seconds=ttl_seconds)
-        ):
-            return trials
-    trials = storage.get_all_trials(study_id, deepcopy=False)
-
-    with in_memory_cache._trials_cache_lock:
-        in_memory_cache._trials_last_fetched_at[study_id] = datetime.now()
-        in_memory_cache._trials_cache[study_id] = trials
-    return trials
-
-
 class InMemoryCache:
     def __init__(self) -> None:
         self._cached_extra_study_property_cache: dict[int, "_CachedExtraStudyProperty"] = {}
@@ -81,6 +50,7 @@ class InMemoryCache:
     def clear(self) -> None:
         self._cached_extra_study_property_cache.clear()
         self._trials_cache.clear()
+        self._trials_last_fetched_at.clear()
 
 
 class _CachedExtraStudyProperty:
