+++ optuna_dashboard/_storage.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime
-import threading
+from datetime import timedelta
 
 from optuna.storages import BaseStorage
 from optuna.storages import RDBStorage
@@ -9,11 +9,36 @@ from optuna.study import StudyDirection
 from optuna.study._frozen import FrozenStudy
 from optuna.trial import FrozenTrial
 
+from ._inmemory_cache import InMemoryCache
 
-# In-memory trials cache
-trials_cache_lock = threading.Lock()
-trials_cache: dict[int, list[FrozenTrial]] = {}
-trials_last_fetched_at: dict[int, datetime] = {}
+
+def get_trials(
+    in_memory_cache: InMemoryCache, storage: BaseStorage, study_id: int
+) -> list[FrozenTrial]:
+    with in_memory_cache._trials_cache_lock:
+        trials = in_memory_cache._trials_cache.get(study_id, None)
+
+        # Not a big fan of the heuristic, but I can't think of anything better.
+        if trials is None or len(trials) < 100:
+            ttl_seconds = 2
+        elif len(trials) < 500:
+            ttl_seconds = 5
+        else:
+            ttl_seconds = 10
+
+        last_fetched_at = in_memory_cache._trials_last_fetched_at.get(study_id, None)
+        if (
+            trials is not None
+            and last_fetched_at is not None
+            and datetime.now() - last_fetched_at < timedelta(seconds=ttl_seconds)
+        ):
+            return trials
+    trials = storage.get_all_trials(study_id, deepcopy=False)
+
+    with in_memory_cache._trials_cache_lock:
+        in_memory_cache._trials_last_fetched_at[study_id] = datetime.now()
+        in_memory_cache._trials_cache[study_id] = trials
+    return trials
 
 
 def get_studies(storage: BaseStorage) -> list[FrozenStudy]:
